controlli di base su inserimento e modifica dati nel db

+++ src/clup/providers/sqlite_admin_provider.py
@@ -16,6 +16,12 @@ class SqliteAdminProvider:
             return admins
 
     def add_admin(self, admin):
+        if admin.id in [a.id for a in self.get_admins()]:
+            raise ValueError("unable to add new admin, id already used")
+
+        if admin.username in [a.username for a in self.get_admins()]:
+            raise ValueError("unable to add new admin, username already used")
+
         with Session(self.engine) as session, session.begin():
             model_admin = models.Admin(
                 uuid=admin.id,
@@ -25,11 +31,17 @@ class SqliteAdminProvider:
             session.add(model_admin)
 
     def remove_admin(self, admin_id):
+        if admin_id not in [a.id for a in self.get_admins()]:
+            raise ValueError("unable to remove admin, id not existing")
+
         with Session(self.engine) as session, session.begin():
             session.query(models.Admin). \
                 filter(models.Admin.uuid == admin_id).delete()
 
     def update_admin(self, admin):
+        if admin.id not in [a.id for a in self.get_admins()]:
+            raise ValueError("unable to update admin, id not existing")
+
         with Session(self.engine) as session, session.begin():
             query = session.query(models.Admin). \
                 filter(models.Admin.uuid == admin.id)
